fix: P2-6 Step1 模型层+计算点 —— 价格字段 Double/double 全量改 BigDecimal
- 实体：StockDailyBar 7 字段、ScreeningMatch 2 字段 Double → BigDecimal
- 模型：KLineData/KLineIterator/StockInfo 17 个价格字段 → BigDecimal（amount 保持 double）
- DTO：5 个（含无调用方的 ScreeningResultDto/ScreeningMatchProjection 随迁）
- GapFillProperties.minPriceThreshold → BigDecimal（YAML 1.00 可绑定）
- 计算/比较点 13 处改 BigDecimal：DataGapFiller（阈值/零价占位/afterHoursChangePercent
  scale 8 divide→setScale 4）、Screening（rise compareTo、lastClose 3 位圆整+去尾零）、
  Tiger/Tiingo StockInfo（divide scale 8 HALF_UP）、ingest 校验与赋值、
  Tiingo/TwelveData RestClient 解析（decimalValue/new BigDecimal）、
  StockDailyBarRepository JPQL 参数 BigDecimal
- 出站序列化：DTO 映射点 stripTrailingZeros（整数值回落普通十进制，避免 1.5E+2 科学计数）
- 转换规则：入站 valueOf(double) 精确表示；Jackson 文本精确解析 BigDecimal

+++ src/main/java/com/stock/invest/config/GapFillProperties.java
@@ -4,6 +4,8 @@ import lombok.Data;
 import org.springframework.boot.context.properties.ConfigurationProperties;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 /**
  * 数据补全相关可配置项。
  */
@@ -15,5 +17,5 @@ public class GapFillProperties {
     /**
      * 最低价格阈值，低于此价格的股票将触发数据补全。
      */
-    private double minPriceThreshold = 1.00;
+    private BigDecimal minPriceThreshold = BigDecimal.valueOf(1.00);
 }
