+++ src/main/java/com/stock/invest/enums/dto/TigerWatchlistRowDto.java
@@ -2,15 +2,17 @@ package com.stock.invest.enums.dto;
 
 import com.fasterxml.jackson.annotation.JsonAlias;
 
+import java.math.BigDecimal;
+
 public record TigerWatchlistRowDto(
     @JsonAlias("code") String symbol,
     String name,
-    @JsonAlias("closePrice") Double lastPrice,
-    Double openPrice,
-    Double highPrice,
-    Double lowPrice,
-    Double changePercent,
-    Double afterHours,
-    Double afterHoursChangePercent,
+    @JsonAlias("closePrice") BigDecimal lastPrice,
+    BigDecimal openPrice,
+    BigDecimal highPrice,
+    BigDecimal lowPrice,
+    BigDecimal changePercent,
+    BigDecimal afterHours,
+    BigDecimal afterHoursChangePercent,
     Object volume
 ) {}
