+++ src/main/java/com/stock/invest/model/StockInfo.java
@@ -3,6 +3,8 @@ package com.stock.invest.model;
 import lombok.Getter;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 /**
  * 股票信息模型
  */
@@ -12,9 +14,9 @@ public class StockInfo {
 
     private String symbol;
     private String name;
-    private double currentPrice;
-    private double openPrice;
+    private BigDecimal currentPrice;
+    private BigDecimal openPrice;
     private long volume;
-    private double change;
-    private double changePercent;
+    private BigDecimal change;
+    private BigDecimal changePercent;
 }
